arm64: dts: renesas: r8a7795: Fix clock, reset and power domain for iVDP1C
According to the R-Car Gen3 Hardware Manual Rev 1.50, the module clock
status, module clock control, and reset control bit of the iVDP1C module
on R-Car H3 ES2.0 or later was changed from bit28 to bit30. And the power
domain of the iVDP1C module was changed from A2VC0 to A2VC1.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>