drm: rcar-du: Fix possible crtcs for plane when using VSPDL
If two or more channels are specified for possible crtc of plane,
if user move the overlay between monitors in the system test,
the following error occurs.

switching CRTC directly, atomic core check failed.

In order to solve the problem, this patch fixes it by fixing
assignment of possible crtc for each plane.

In addtion, add an option to change the assignment of the
overlay sharing with VSPDL.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>