clk: renesas: rcar-gen3: Fix SCCG/Clean peripheral clocks definition
The naming in SCCG/Clean peripheral clocks definition should be defined
in the other way. This change aligns its definition with Mode pin clock
selection definition which it inherits from, the detail clock definition
in <*-cpg-mssr.c>, the CLK_TYPE_GEN3_MDSEL handling in rcar-gen3-cpg.c
and the datasheet.

Signed-off-by: Hai Nguyen Pham <hai.pham.ud@renesas.com>