rcar-vin: add support for UDS (Up Down Scaler)
This series adds support for Renesas R-Car Gen3 VIN Up Down Scaler
(UDS). Some VIN instances have access to a often shared UDS which can be
used to scale the captured image up or down. If the scaler is shared it
can only be used exclusively by one VIN at a time, switching in runtime
and detection if a UDS are in use is supported in this series. If the
user tries to start a capture on a VIN which would require the use of a
scaler but that scaler is in use -EBUSY is returned.

Patch 1/3 fix a format alignment issue found when working with UDS
support. While patch 2/3 ands the UDS logic and 3/3 defines which VIN on
which SoC have access to a UDS and how it's shared.

The series is based on top of media-tree/master and is tested on R-Car
Gen3 H3, M3-W, M3-N and Gen2 Koelsch (checking for regressions as Gen2
have no UDS).

* Changes since v1
- Patch 1/3 have been replaced with a less strict version after good
  comments from Hans.

Niklas S旦derlund (3):
  rcar-vin: align width before stream start
  rcar-vin: add support for UDS (Up Down Scaler)
  rcar-vin: declare which VINs can use a Up Down Scaler (UDS)

This patch based on https://patchwork.kernel.org/project/linux-media/cover/20181004200402.15113-1-niklas.soderlund+renesas@ragnatech.se/

Fixed to compare with compose.width instead of format.width.

Signed-off-by: Niklas Soderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>