net: ethernet: renesas: rswitch: Fix SerDes initialization
After completing the trigger of the reset on R-CarS4 v1.1,
it is necessary to clear the initial value of the LOCAL REGISTER
to ensure the stability of initialization. Otherwise, there are
ratio errors when using USXGMII mode.

Additionally, using retry five times to restart SerDes link up
in case initialization fails.

Signed-off-by: Tam Nguyen <tam.nguyen.xa@renesas.com>
Signed-off-by: Phong Hoang <phong.hoang.wz@renesas.com>