soc: renesas: rcar-sysc: Add H3 for clock control
This patch supports clock control for H3 versions: 1.1, 2.0, 3.0.

Currently, the VSPM and VPSD drivers don't use the Runtime PM APIs
fully to control clocks of modules. That's why those modules will
be in the no_off_mask field to prevent conflicting as described in
the commit (Add no_off_mask field for clock control).
The clocks of those modules in H3 are controlled by the SMSTPCR6
register.
Those modules are:
- fcpvd{0, 1, 2}
- fcpvb{0, 1}
- fcpvi{0, 1}
- vspd{0, 1, 2}
- vspbc
- vspbd
- vspi{0, 1}

Signed-off-by: Vinh Nguyen <vinh.nguyen.xz@renesas.com>