drm: rcar-du: lvds: in hdmi-dual mode, after wake-up, reset both LVDS0, LVDS1 after wake-up
On vga-hdmi-dual mode, DU1 (VGA) used dotclock LVDS1.
When screen-saver event happen, DU1 was reset, so LVDS1 reset.
But LVDS0 was not reset, because when screen-saver, DO NOT call rcar_lvds_atomic_disable(), and when wake-up, only call rcar_lvds_atomic_enable().
So, HDMI wake-up failed.

To avoid this problem, on vga-hdmi-dual mode,
  when calling rcar_lvds_atomic_enable(), call reset_control_reset() instead of call reset_control_deassert().
  when calling rcar_lvds_atomic_disable(), ignore call reset_control_assert().

Signed-off-by: Nguyen Bao Nguyen <nguyen.nguyen.yj@renesas.com>